requirements.txt
django==1.11.6
django-allauth==0.33.0
pillow==4.3.0
gunicorn==19.7.1
django-braces==1.12.0
psycopg2==2.7.4
raven==6.8.0
celery==4.1.0
flower==0.9.2
redis==2.10.6
kombu==4.1.0
invoke==0.16.3
